feat: add linters, change css to scss, change project structure
