+++ .stylelintrc.json
@@ -1,0 +1,3 @@
+{
+  "extends": ["stylelint-config-standard-scss", "stylelint-config-prettier-scss"]
+}
