--- goods.html
@@ -1,81 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta http-equiv="X-UA-Compatible" content="IE=edge">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <link rel="preconnect" href="https://fonts.googleapis.com">
-    <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
-    <link href="https://fonts.googleapis.com/css2?family=Lora&display=swap" rel="stylesheet">
-    <title>Online-store</title>
-</head>
-<body>
-    <header class="header">
-        <div class="wrapper wrapper-header"></div>
-    </header>
-    <main class="main">
-      <section class="card">
-        <div class="card_info">
-          <h3 class="card-title">Настольная голубая елка в мешочке Верона 60 см</h3>
-          <div class="card-column">
-              <div class="card_image">
-                <img class="card_image_big" src="./id1.jpg" alt="goods">
-                <div class="card_image_small">
-                  <img class="card_image_small_1" src="./id1.jpg" alt="goods">
-                  <img class="card_image_small_2" src="./id1_1.jpg" alt="goods">
-                </div>
-              </div>
-              <div class="card_details">
-                <div class="card_brand">
-                  <div>Брэнд: </div>
-                  <div class="card_brand_details">Beatrees</div>
-                </div>
-                <div class="card_category">
-                  <div>Категория: </div>
-                  <div class="card_category_details">Искусственные елки</div>
-                </div>
-                <div class="card_cost">
-                  <div>Цена (в рублях): </div>
-                  <div class="card_cost_details">2030</div>
-                </div>
-                <div class="card_quantity">
-                  <div>Количество на складе: </div>
-                  <div class="card_quantity_details">7</div>
-                </div>
-                <div class="card_description">
-                  <div>Описание: </div>
-                  <div class="card_description_details">Ель закреплена на устойчивой подставке в виде декоративного мешочка, которая позволяет расположить ее на столе, подоконнике или в любом удобном для вас месте. Эта модель елки не требует сборки, достаточно просто достать ее из упаковки и слегка распушить ветви - и ваша новогодняя красавица готова радовать вас и ваших близких!</div>
-                </div>
-              </div>
-          </div>
-          <div class="info-btn">
-              <button class="btn btn-cart">Добавить в корзину</button>
-          </div>
-      </div>
-
-      </section>
-
-    </main>
-    <footer class="footer">
-        <div class="wrapper wrapper-footer">
-            <ul class="footer_list">
-                <li class="footer_list_item">
-                    <a href="https://github.com/nbelchikova">Natalia</a>
-                </li>
-                <li class="footer_list_item">
-                  <a href="https://github.com/Kunitsa3">Olga</a>
-              </li>
-                <li class="footer_list_item">
-                    <a href="">2022</a>
-                </li>
-                <li class="footer_list_item">
-                    <img src="./rs.png" alt="logo">
-                </li>
-                <li class="footer_list_item">
-                    <a href="https://rs.school/js/">Link</a>
-                </li>
-            </ul>
-        </div>
-    </footer>
-</body>
-</html>
