+++ package.json
@@ -4,13 +4,17 @@
   "description": "",
   "main": "webpack.base.config.js",
   "scripts": {
-      "start": "webpack serve --open --config ./webpack.config.js --env mode=dev",
-      "build": "webpack --config ./webpack.config.js --env mode=prod",
-      "lint": "eslint . --ext .ts"
+    "start": "webpack serve --open --config ./webpack.config.js --env mode=dev",
+    "build": "webpack --config ./webpack.config.js --env mode=prod",
+    "lint": "eslint . --ext .ts",
+    "lint:fix": "eslint --fix . --ext .ts",
+    "prettier": "prettier --write .",
+    "stylelint": "stylelint \"**/*.scss\" --fix",
+    "fix": "npm-run-all lint:fix prettier stylelint"
   },
   "repository": {
-      "type": "git",
-      "url": ""
+    "type": "git",
+    "url": ""
   },
   "author": "Kris Casper",
   "devDependencies": {
@@ -21,12 +25,19 @@
     "css-loader": "^5.2.4",
     "eslint": "^7.25.0",
     "eslint-config-airbnb-typescript": "^12.3.1",
+    "eslint-config-prettier": "^8.5.0",
     "eslint-plugin-import": "^2.22.1",
+    "eslint-plugin-prettier": "^4.2.1",
     "eslint-webpack-plugin": "^2.5.4",
     "html-webpack-plugin": "^5.3.1",
     "mini-css-extract-plugin": "^1.6.0",
+    "npm-run-all": "^4.1.5",
+    "prettier": "2.8.1",
     "sass": "^1.32.12",
     "sass-loader": "^11.0.1",
+    "stylelint": "^14.16.0",
+    "stylelint-config-prettier-scss": "^0.0.1",
+    "stylelint-config-standard-scss": "^6.1.0",
     "ts-loader": "^9.1.1",
     "typescript": "^4.2.4",
     "webpack": "^5.36.2",
