--- src/goods.ts
@@ -1,4 +1,0 @@
-import './reset.css';
-import './index.css';
-import './goods.css';
-
