+++ src/pages/cart/shoppingCart.html
@@ -1,0 +1,39 @@
+<!DOCTYPE html>
+<html lang="en">
+  <head>
+    <meta charset="UTF-8" />
+    <meta http-equiv="X-UA-Compatible" content="IE=edge" />
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+    <link rel="preconnect" href="https://fonts.googleapis.com" />
+    <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin />
+    <link href="https://fonts.googleapis.com/css2?family=Lora&display=swap" rel="stylesheet" />
+    <title>Online-store</title>
+  </head>
+  <body>
+    <header class="header">
+      <div class="wrapper wrapper-header"></div>
+    </header>
+    <main class="main"></main>
+    <footer class="footer">
+      <div class="wrapper wrapper-footer">
+        <ul class="footer_list">
+          <li class="footer_list_item">
+            <a href="https://github.com/nbelchikova">Natalia</a>
+          </li>
+          <li class="footer_list_item">
+            <a href="https://github.com/Kunitsa3">Olga</a>
+          </li>
+          <li class="footer_list_item">
+            <a href="">2022</a>
+          </li>
+          <li class="footer_list_item">
+            <img src="./rs.png" alt="logo" />
+          </li>
+          <li class="footer_list_item">
+            <a href="https://rs.school/js/">Link</a>
+          </li>
+        </ul>
+      </div>
+    </footer>
+  </body>
+</html>
