+++ src/pages/cart/shoppingCart.scss
@@ -1,0 +1,1 @@
+@import '../../styles/index';
