+++ src/pages/cart/shoppingCart.ts
@@ -1,0 +1,2 @@
+import '../../styles/index.scss';
+import './shoppingCart.scss';
