+++ src/pages/goods/goods.scss
@@ -1,61 +1,69 @@
+@import '../../styles/index';
 
 .card_info {
   display: block;
-  margin: 110px auto 110px auto;
+  margin: 110px auto;
   font-size: 15px;
-  border: #36B0CB 2px solid;
+  border: 2px solid $main;
   border-radius: 10px;
   padding-left: 20px;
   padding-right: 20px;
   padding-top: 20px;
   padding-bottom: 20px;
 }
+
 .card-title {
   font-size: 25px;
-  color: #36B0CB;
+  color: $main;
   text-decoration: underline;
   text-align: center;
   margin-bottom: 30px;
 }
+
 .card-column {
   display: flex;
 }
+
 .card_image {
   width: 48%;
   display: block;
-
 }
-.card_image_big{
+
+.card_image_big {
   width: 400px;
-
 }
+
 .card_details {
   width: 48%;
   display: block;
 }
+
 .card_image_small {
   display: flex;
   width: 400px;
   justify-content: space-around;
   margin-top: 30px;
 }
+
 .card_image_small_1 {
   width: 150px;
-
 }
+
 .card_image_small_2 {
   width: 150px;
 }
+
 .card_details {
-  color: #36B0CB;
+  color: $main;
   font-size: 20px;
 }
+
 .card_brand_details,
 .card_category_details,
 .card_cost_details,
 .card_quantity_details,
 .card_description_details {
-  color: black;
+  color: $black;
   font-size: 15px;
   margin-bottom: 30px;
 }
