+++ src/pages/goods/goods.ts
@@ -1,0 +1,1 @@
+import './goods.scss';
