+++ src/pages/homepage/index.scss
@@ -1,55 +1,40 @@
+@import '../../styles/index';
+
 html {
-  background-image: url('../public/background.jpg');
+  background-image: url('../../../public/background.jpg');
   background-repeat: repeat;
   background-position: center;
   background-size: cover;
   font-size: 20px;
   box-sizing: border-box;
-  font-family: 'Lora', serif;
+  font-family: Lora, serif;
 }
+
 .wrapper {
   max-width: 1100px;
   margin: 0 auto;
 }
-.header, .footer {
+
+.header {
   display: block;
-  background-color: #36B0CB;
+  background-color: $main;
   height: 80px;
   border-radius: 10px;
   width: 100%;
-}
-.header {
-  position:fixed;
-  top:0px;
+  position: fixed;
+  top: 0;
   z-index: 5;
-}
-.footer {
-  position:fixed;
-  bottom:0px;
-  z-index: 5;
-}
-
-.footer_list {
-  display: flex;
-  height: 80px;
-  justify-content: space-around;
-  align-items:center;
-}
-.footer_list_item a {
-  color: white;
-}
-.footer_list_item img {
-  width:50px;
 }
 
 .slider {
   margin-top: 100px;
   position: relative;
 }
+
 .slider_items_img {
   display: none;
   width: 1000px;
-  margin: 20px auto 50px auto;
+  margin: 20px auto 50px;
 }
 
 .array-left {
@@ -58,89 +43,101 @@ html {
   top: 120px;
   z-index: 10;
 }
+
 .array-right {
   position: absolute;
   right: 0;
   top: 120px;
   z-index: 10;
 }
+
 .shop {
   display: flex;
   margin: 0 auto;
   margin-bottom: 100px;
 }
+
 .shop_filter {
   width: 370px;
-  border: #36B0CB 2px solid;
+  border: 2px solid $main;
   height: 500px;
   margin-right: 30px;
   border-radius: 10px;
   padding-left: 20px;
 }
+
 .shop_items {
   width: 700px;
-  border: #36B0CB 2px solid;
+  border: 2px solid $main;
   height: auto;
   border-radius: 10px;
 }
+
 .shop_filter_title {
   font-size: 20px;
-  color:#36B0CB;
+  color: $main;
   text-align: center;
   margin-bottom: 10px;
 }
+
 .p_filter {
   margin-bottom: 10px;
   margin-top: 10px;
 }
+
 .btn {
-  margin: 30px auto 30px auto;
-  background-color: #36B0CB;
-  color: white;
+  margin: 30px auto;
+  background-color: $main;
+  color: $white;
   text-align: center;
-  border: #36B0CB;
+  border: $main;
   height: 20px;
 }
+
 .cost-filter {
   margin-left: 20px;
 }
+
 .form {
   display: flex;
   flex-direction: column;
   font-size: 15px;
 }
+
 .good-items {
   margin-top: 30px;
   display: flex;
-  width:700px;
+  width: 700px;
   flex-wrap: wrap;
 }
+
 .good-items_info {
   width: 326px;
-  border: #36B0CB 2px solid;
+  border: 2px solid $main;
   height: 300px;
   margin-right: 10px;
   margin-left: 10px;
   margin-bottom: 10px;
 }
+
 .info-title {
   text-align: center;
   text-decoration: underline;
-  color: #36B0CB;
+  color: $main;
 }
+
 .info-column {
   display: flex;
   margin-top: 20px;
   font-size: 15px;
-
 }
+
 .info-column_image img {
   width: 150px;
   margin-right: 20px;
   margin-left: 10px;
 }
+
 .info-btn {
   display: flex;
 }
-
-
