--- src/shoppingCart.css
@@ -1,1 +1,0 @@
-
