--- src/shoppingCart.ts
@@ -1,3 +1,0 @@
-import './reset.css';
-import './index.css';
-import './shoppingCart.css';
