+++ src/styles/colors.scss
@@ -1,0 +1,3 @@
+$main: #36b0cb;
+$white: #fff;
+$black: #000;
