+++ src/styles/index.scss
@@ -1,0 +1,3 @@
+@import './reset';
+@import './colors';
+@import './footer';
