+++ webpack.dev.config.js
@@ -1,13 +1,13 @@
 const path = require('path');
 
 module.exports = {
-    mode: 'development',
-    devtool: 'inline-source-map',
-    devServer: {
-        static: {
-            directory: path.resolve(__dirname, '../dist'),
-        },
-        compress: true,
-        port: 9000,
+  mode: 'development',
+  devtool: 'inline-source-map',
+  devServer: {
+    static: {
+      directory: path.resolve(__dirname, '../dist'),
     },
+    compress: true,
+    port: 9000,
+  },
 };
