+++ webpack.prod.config.js
@@ -1,3 +1,3 @@
 module.exports = {
-    mode: 'production',
+  mode: 'production',
 };
